backend/requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
supabase==2.0.3
instaloader==4.10
openai-whisper==20231117
groq==0.4.1
TikTokApi==6.0.0
python-dotenv==1.0.0
requests==2.31.0
playwright==1.40.0
